Finished up some basic attack logic
Added an offset to the constructor so that a player on the right side has a hit box that goes in the direction of the player on the left side. Broke out collision detection if statement into it's own function. The new function takes an object with a key of rectangles or hitboxes and a value of player or enemy (player 1 or player 2). Lastly added an event listener for player 2 to hit player 1